Other public/private adjustments along with switching to api autodocs
