--- docs/api.md
@@ -1,37 +1,0 @@
-# API Docs
-
-## Networkx Backend
-
-::: geff.read_nx
-
-::: geff.write_nx
-
-## Utilities
-
-::: geff.validate
-
-## Metadata
-
-::: geff.GeffMetadata
-
-::: geff.metadata_schema.Axis
-
-::: geff.metadata_schema.axes_from_lists
-
-## Units
-
-::: geff.valid_values.validate_axis_type
-
-::: geff.valid_values.VALID_AXIS_TYPES
-
-::: geff.valid_values.validate_space_unit
-
-::: geff.valid_values.VALID_SPACE_UNITS
-
-::: geff.valid_values.validate_time_unit
-
-::: geff.valid_values.VALID_TIME_UNITS
-
-## Writing Helpers
-
-::: geff.write_arrays.write_arrays
