+++ mkdocs.yml
@@ -9,7 +9,6 @@ nav:
   - what-is-geff.md
   - specification.md
   - command-line-tools.md
-  - api.md
 
 theme:
   name: "material"
@@ -62,3 +61,5 @@ plugins:
           # summary: true
           line_length: 60
 - include-markdown
+- api-autonav:
+    modules: ['src/geff']
