+++ src/geff/_graph_libs/_rustworkx.py
@@ -16,8 +16,8 @@ except ImportError as e:
 
 from geff.core_io import write_dicts
 from geff.core_io._base_read import read_to_memory
-from geff.core_io.utils import calculate_roi_from_nodes
-from geff.metadata.schema import GeffMetadata, _axes_from_lists
+from geff.core_io._utils import calculate_roi_from_nodes
+from geff.metadata._schema import GeffMetadata, _axes_from_lists
 from geff.metadata.utils import (
     create_or_update_metadata,
     get_graph_existing_metadata,
@@ -27,7 +27,7 @@ if TYPE_CHECKING:
     from numpy.typing import NDArray
     from zarr.storage import StoreLike
 
-    from geff.typing import PropDictNpArray
+    from geff._typing import PropDictNpArray
 
 
 def get_roi_rx(
