+++ src/geff/convert/_ctc.py
@@ -19,7 +19,7 @@ from zarr.storage import StoreLike
 
 import geff
 from geff.core_io import write_arrays
-from geff.metadata.schema import Axis, GeffMetadata
+from geff.metadata._schema import Axis, GeffMetadata
 
 
 def ctc_tiffs_to_zarr(
