+++ src/geff/core_io/_base_write.py
@@ -7,7 +7,7 @@ import numpy as np
 import zarr
 
 from geff import _path
-from geff.core_io.utils import remove_tilde
+from geff.core_io._utils import remove_tilde
 from geff.metadata._valid_values import validate_data_type
 
 if TYPE_CHECKING:
@@ -15,7 +15,7 @@ if TYPE_CHECKING:
 
     from zarr.storage import StoreLike
 
-    from geff.metadata.schema import GeffMetadata
+    from geff.metadata._schema import GeffMetadata
 
 
 def write_dicts(
