+++ src/geff/metadata/__init__.py
@@ -1,0 +1,3 @@
+from ._schema import Axis, DisplayHint, GeffMetadata, PropMetadata, RelatedObject
+
+__all__ = ["Axis", "DisplayHint", "GeffMetadata", "PropMetadata", "RelatedObject"]
