+++ tests/test_geff_reader.py
@@ -1,8 +1,8 @@
 import numpy as np
 import pytest
 
-from geff.core_io import GeffReader
-from geff.graph_libs._networkx import construct_nx
+from geff import GeffReader
+from geff._graph_libs._networkx import construct_nx
 from geff.testing.data import create_memory_mock_geff
 
 node_id_dtypes = ["int8", "uint8", "int16", "uint16"]
