+++ tests/test_rustworkx/test_rx_basics.py
@@ -3,7 +3,7 @@ import pytest
 import zarr
 
 import geff
-from geff.metadata.schema import GeffMetadata, _axes_from_lists
+from geff.metadata._schema import GeffMetadata, _axes_from_lists
 from geff.testing.data import create_memory_mock_geff
 
 rx = pytest.importorskip("rustworkx")
